Changed the syntax of add and use function first and created handler for non-existing branches on use function
Instead of calling `hit branch <command>` you have to call `hit <branch_command>` instead now if you want to access branch features. There will be methods for 'git remote' in the feature but because branches are the main focus of hit, they don't need a 'branch' keyword anymore. Remote- and other features will be callable with a prefix like 'hit remote <remote_command>'  when they are added in the future.